backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-storages==1.14.2
django-redis==5.4.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
celery==5.3.6
redis==5.0.1
stripe==7.12.0
Pillow==10.2.0
boto3==1.34.25
python-decouple==3.8
whitenoise==6.6.0
python-magic==0.4.27
hashids==1.3.1
shortuuid==1.0.11
